code refactor

+++ app/src/main/java/com/example/todolist/TaskActivity.kt
@@ -7,7 +7,6 @@ import android.content.Context
 import android.graphics.Color
 import android.os.Bundle
 import android.text.TextUtils
-import android.util.Log
 import android.view.View
 import android.view.ViewGroup
 import android.view.inputmethod.InputMethodManager
@@ -79,15 +78,10 @@ class TaskActivity : AppCompatActivity(), View.OnClickListener {
         val category2 = intent.getStringExtra("category").toString()
         val date2 = intent.getStringExtra("date")
         val time2 = intent.getStringExtra("time")
-        val position2 = intent.getIntExtra("position",-1)
-        Log.e("ok2",position2.toString())
-
-
 
         if(title2!=null) {
             binding.inputTitle.setText(title2.toString())
             binding.toolbarTaskTitle.text = "Edit Task"
-            // below is to set cursor position of title at end of the textview
             binding.inputTitle.setSelection(binding.inputTitle.text.toString().length)
         }
         val defaultDate:String
@@ -115,8 +109,6 @@ class TaskActivity : AppCompatActivity(), View.OnClickListener {
         date3 = date2 ?: defaultDate
         time3 = time2 ?: defaultTime
 
-        // how the heck people do android dev? it is literally trash
-
         val spinnerAdapter = binding.spinnerCategory.adapter as ArrayAdapter<String>
         val pos = spinnerAdapter.getPosition(category2)
         binding.spinnerCategory.setSelection(Math.max(0,pos))
@@ -124,9 +116,7 @@ class TaskActivity : AppCompatActivity(), View.OnClickListener {
         binding.addNewCategoryBtn.setOnClickListener {
             addNewCategory()
         }
-
     }
-
 
     private fun addNewCategory(){
         val dialog = Dialog(binding.inputTitle.context)
@@ -135,8 +125,11 @@ class TaskActivity : AppCompatActivity(), View.OnClickListener {
             parent.removeView(bindingAddCategory.root)
         }
         dialog.setContentView(bindingAddCategory.root)
+        val w = resources.displayMetrics.widthPixels*0.95   // to occupy 95% of screen's width
+        dialog.window?.setLayout(w.toInt(),ViewGroup.LayoutParams.WRAP_CONTENT)
         dialog.show()
         bindingAddCategory.newCategoryOk.isEnabled = false
+        bindingAddCategory.newCategoryName.text.clear()
         bindingAddCategory.newCategoryOk.setTextColor(getColor(R.color.gray))
         bindingAddCategory.newCategoryName.addTextChangedListener {
             var input = bindingAddCategory.newCategoryName.text.toString()
@@ -154,37 +147,29 @@ class TaskActivity : AppCompatActivity(), View.OnClickListener {
         bindingAddCategory.newCategoryOk.setOnClickListener {
             val newCategory = bindingAddCategory.newCategoryName.text.toString()
             newCategory.trim()
-            var categoryAlreadyExists = false
-            runBlocking {
-                launch(Dispatchers.IO) {
-                    if(newCategory in categoryDatabase.categoryDao().getCategoriesNames())
-                        categoryAlreadyExists = true
-                }
-            }
-            if (categoryAlreadyExists){
-                val imm = getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
-                imm.toggleSoftInput(InputMethodManager.HIDE_IMPLICIT_ONLY, 0)
-                val snack = Snackbar.make(binding.inputTitle,"Category already exists!",Snackbar.LENGTH_SHORT)
-                snack.show()
-                dialog.dismiss()
-            }
-            else{
-                GlobalScope.launch(Dispatchers.Main){
-                    val id = withContext(Dispatchers.IO) {
-                        return@withContext categoryDatabase.categoryDao().insertCategory(
-                            CategoryModel(
-                                newCategory
-                            )
-                        )
+
+            lifecycleScope.launch(Dispatchers.IO) {
+                val categoryNames = categoryDatabase.categoryDao().getCategoriesNames()
+                if(newCategory in categoryNames) {
+                    val job = async(Dispatchers.Main) {
+                        val imm = getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
+                        imm.toggleSoftInput(InputMethodManager.HIDE_IMPLICIT_ONLY, 0)
+                        val snack = Snackbar.make(binding.inputTitle,"Category already exists!",Snackbar.LENGTH_SHORT)
+                        snack.show()
                     }
+                    job.await()
                 }
-                val spinnerAdapter = binding.spinnerCategory.adapter as ArrayAdapter<String>
-                spinnerAdapter.add(newCategory)
-
-                spinnerAdapter.notifyDataSetChanged()
-                dialog.dismiss()
+                else {
+                    categoryDatabase.categoryDao().insertCategory(CategoryModel(newCategory))
+                    val job = async(Dispatchers.Main) {
+                        val spinnerAdapter = binding.spinnerCategory.adapter as ArrayAdapter<String>
+                        spinnerAdapter.add(newCategory)
+                        spinnerAdapter.notifyDataSetChanged()
+                    }
+                    job.await()
+                }
             }
-
+            dialog.dismiss()
         }
     }
 
@@ -207,7 +192,6 @@ class TaskActivity : AppCompatActivity(), View.OnClickListener {
         if(title2!=null)
             dialogText = "Discard changes?"
         bindingDiscard.discard.text = dialogText
-//        yes_button.setBackgroundColor(Color.RED)  // why this doesn't happen???
         bindingDiscard.cancelButton.setOnClickListener {
             dialog.dismiss()
         }
@@ -260,8 +244,8 @@ class TaskActivity : AppCompatActivity(), View.OnClickListener {
                 setTimeListener()
             }
             R.id.save_todo -> {
-                val position2 = intent.getIntExtra("position",-1)
-                saveTodo(position2)
+                val position = intent.getIntExtra("position",-1)
+                saveTodo(position)
             }
         }
     }
@@ -274,32 +258,15 @@ class TaskActivity : AppCompatActivity(), View.OnClickListener {
             Toast.makeText(binding.inputTitle.context,"Please enter a title",Toast.LENGTH_SHORT).show()
         else {
             var isUpdated = false
-            if (position!=-1){
+            if (position!=-1) {
                 isUpdated = true
-                GlobalScope.launch(Dispatchers.IO) {
-                    todoDatabase.todoDao().deleteTask(position.toLong())
-                }
             }
-//            val myFormat = "EEE, d MMM yyyy h:mm a"
-//            val sdf = SimpleDateFormat(myFormat)
-//            val myFormat1 = "EEE, d MMM yyyy"
-//            val sdf1 = SimpleDateFormat(myFormat1)
-//            val myFormat2 = "h:mm a"
-//            val sdf2 = SimpleDateFormat(myFormat2)
-//            val currentDateTime = sdf.format(Date(System.currentTimeMillis()))
-//            Log.e("compare 1","${sdf.format(Date(myCalendar.timeInMillis))}: ${currentDateTime} and ${myCalendar.timeInMillis}: ${System.currentTimeMillis()}")
-//            Log.e("compare 2","${sdf.format(Date(finalDate))} and ${sdf.format(Date(finalTime))}")
-//            Log.e("compare 3","${sdf1.format(Date(finalDate))} and ${sdf2.format(Date(finalTime))}")
-//            val combinedDateTime = sdf1.format(finalDate)+" "+ sdf2.format(finalTime)
-//            if (sdf.parse(combinedDateTime).before(sdf.parse(currentDateTime)))
-//                Log.e("compare 4"," ${combinedDateTime} is before ${currentDateTime}")
-//            else if(sdf.parse(combinedDateTime).after(sdf.parse(currentDateTime)))
-//                Log.e("compare 4"," ${combinedDateTime} is after ${currentDateTime}")
-//            else
-//                Log.e("compare 4"," ${combinedDateTime} is equal to ${currentDateTime}")
-            GlobalScope.launch(Dispatchers.Main){
-                val id = withContext(Dispatchers.IO) {
-                    return@withContext todoDatabase.todoDao().insertTask(
+            lifecycleScope.launch(Dispatchers.IO) {
+                val job = async {
+                    if (isUpdated) {
+                        todoDatabase.todoDao().deleteTask(position.toLong())
+                    }
+                    todoDatabase.todoDao().insertTask(
                         TodoModel(
                             title,
                             description,
@@ -309,11 +276,15 @@ class TaskActivity : AppCompatActivity(), View.OnClickListener {
                         )
                     )
                 }
+                job.await()
+                withContext(Dispatchers.Main) {
+                    var msg = "To Do saved"
+                    if (isUpdated)
+                        msg = "To Do updated"
+                    Toast.makeText(this@TaskActivity,msg,Toast.LENGTH_SHORT).show()
+                    finish()
+                }
             }
-            val msg = if (isUpdated) "To Do saved" else "To Do updated"
-            Snackbar.make(findViewById(android.R.id.content),msg,Snackbar.LENGTH_SHORT).show()
-            Toast.makeText(this@TaskActivity,msg,Toast.LENGTH_SHORT).show()
-            finish()
         }
     }
 
@@ -321,8 +292,6 @@ class TaskActivity : AppCompatActivity(), View.OnClickListener {
     fun finishTask(view:View){
         onBackPressed()
     }
-
-
 
 
     private fun setTimeListener() {
@@ -385,7 +354,6 @@ class TaskActivity : AppCompatActivity(), View.OnClickListener {
     }
 
     private fun updateDate2(currDate:String) {
-        // ADD THIS FUNCTION TO A CLASS
         //Mon, 5 Jan 2020
         myCalendar = Calendar.getInstance()
         val n = currDate.length
@@ -423,9 +391,7 @@ class TaskActivity : AppCompatActivity(), View.OnClickListener {
 
     private fun updateTime2(currTime:String) {
 
-        // ADD THIS FUNCTION TO A CLASS
         //Mon, 5 Jan 2020
-
         myCalendar = Calendar.getInstance()
         val n = currTime.length
         if(n==8){
@@ -450,8 +416,5 @@ class TaskActivity : AppCompatActivity(), View.OnClickListener {
         val sdf = SimpleDateFormat(myFormat)
         finalTime = myCalendar.time.time
         binding.inputTime.setText(sdf.format(myCalendar.time))
-
     }
-
-
 }
